+++ tests/helpers/helpers.js
@@ -1,10 +1,31 @@
 import { skip, test } from 'qunit';
 import { gte } from 'ember-compatibility-helpers';
 import Ember from 'ember';
+import {
+  importSync,
+  macroCondition,
+  dependencySatisfies,
+} from '@embroider/macros';
 
 export const decoratorTest = gte('3.10.0') ? test : skip;
 
 export function makeAsyncError(hooks) {
   hooks.afterEach(() => (Ember.onerror = null));
   return () => new window.Promise((r) => (Ember.onerror = r));
+}
+
+export function getDebugFunction(type) {
+  if (macroCondition(dependencySatisfies('ember-source', '>3.26.0'))) {
+    return importSync('@ember/debug').getDebugFunction(type);
+  } else {
+    return Ember[type];
+  }
+}
+
+export function setDebugFunction(type, fn) {
+  if (macroCondition(dependencySatisfies('ember-source', '>3.26.0'))) {
+    return importSync('@ember/debug').setDebugFunction(type, fn);
+  } else {
+    Ember[type] = fn;
+  }
 }